Add metadata indicating if a type is resizable
A resizable desktop type is one in which `xrandr` can be used to
*sensibly* adjust the size of the desktop session.  Here sensibly refers
to the session being able to adjust itself to the new size correctly.

An example of a non-resizable type is `chrome`.  Whilst the VNC session
is correctly resized, the chrome browser is not.  Either it sits in the
top left hand corner failing to take up all of the available space or it
spills over the available desktop space and parts of it become
inaccessible.

It would perhaps be possible to extend the geometry script so that the
chrome browser is resized to match the available space, but that is left
as a future possible enhancement.